+ Organ, RotarySpeaker: moved to "main" group (no longer needs --enable-experimental) + Organ: added percussion level (in dB) + Organ: made percussion decay time freely adjustable + Knob: draw focus rectangle properly + JACK Host: better plugin naming on buttons in main window
git-svn-id: https://calf.svn.sourceforge.net/svnroot/calf/trunk@132 78b06b96-2940-0410-b7fc-879d825d01d8

+++ src/synth.cpp
@@ -126,7 +126,7 @@ void basic_synth::control_change(int ctl, int val)
             control_change(64, 0);
         for (int i = 0; i < keystack.count(); i++)
             notes.push_back(keystack.nth(i));
-        for (int i = 0; i < notes.size(); i++)
+        for (int i = 0; i < (int)notes.size(); i++)
             note_off(notes[i], 0);
     }
     if (ctl == 121) {
